+++ llvm/lib/Target/C166/C166ISelLowering.h
@@ -59,6 +59,9 @@ public:
                              SelectionDAG &DAG) const;
   void ReplaceSignedDivBy16Results(SDNode *N, SmallVectorImpl<SDValue> &Results,
                                    SelectionDAG &DAG) const;
+  void ReplaceMinMaxResults(SDNode *N, SmallVectorImpl<SDValue> &Results,
+                            SelectionDAG &DAG) const;
+
   void ReplaceNodeResults(SDNode *N, SmallVectorImpl<SDValue> &Results,
                           SelectionDAG &DAG) const override;